lightwight blocking, change ADC from 10 to 8 bit resolution

+++ kernel/tasks.h
@@ -45,7 +45,7 @@ struct rgbarg {
 	uint8_t r;
 	uint8_t g;
 	uint8_t b;
-	uint16_t *v;
+	uint8_t *v;
 };
 
 struct pwmarg {
